Includes brainwallet code
Description: Calling the wallet with the brainwallet flag will prompt the user to type in his brainwallet phrase, which will be hashed with cn_fast_hash. The output is used as the hexadecimal seed.

Disclaimer: I don't know what I am doing :) I am just testing the waters with Monero, C++ and Github with what seemed to be the simplest thing to implement. The code works (as far as I can tell) but I am looking forward to hearing suggestions on what to improve or do differently!